Move school meal variables into their own files
